clean up themes data handling
