+++ src/storage/custom-css.js
@@ -1,12 +1,8 @@
 const path = require("path")
 const { remote } = require("electron")
-const Bus = require("../bus")
 
 const config_dir = remote.app.getPath("userData")
 const customCSS = path.join(config_dir, "user.css")
-
-const THEME_ORIGINAL = "original"
-const THEME_DARK_KING = "dark-king"
 
 exports.customCSS = customCSS
 
@@ -29,14 +25,3 @@ exports.injectCSS = function (fileName = customCSS) {
     head.appendChild(link)
   })
 }
-
-Bus.on(Bus.events.CHANGE_THEME, (data) => {
-  const stylesheet = document.querySelector(
-    "#theme-stylesheet"
-  )
-  if (data.theme === THEME_ORIGINAL) {
-    stylesheet.href = "./app/styles/themes/original.css"
-  } else if (data.theme === THEME_DARK_KING) {
-    stylesheet.href = "./app/styles/themes/dark-king.css"
-  }
-})
