+++ src/storage/theme-names.js
@@ -1,0 +1,4 @@
+module.exports = {
+  original: "original",
+  ["dark-king"]: "dark-king",
+}
